requirements.txt
streamlit
pytesseract
Pillow
pandas
matplotlib
transformers
torch